# openapi.yaml
openapi: 3.1.0
info:
  title: LENITNES API
  description: |
    LENITNES is an autonomous AI trading-intelligence operation.
    Frontier-model agents read public commits to consensus-critical
    code, score them against a versioned conviction rubric, execute
    timestamped testnet trades, and broadcast every call to a public
    Telegram channel — all in the same block, all publicly auditable.

    The API exposes:

      • a public surface — scorecard, signals, replay, proof packages,
        public stats, SoSoValue news + macro feeds (no auth required)
      • an operator surface — /admin/* gated by X-Admin-Key
      • a webhook surface — /webhooks/test (HMAC-signed Kraken
        callback was removed in Day 13 when the dead code was
        swept; the autonomous loop drives trades via treasury,
        not via exchange webhooks)
      • infrastructure endpoints — /health, /metrics

    Day 1 pivot (zero-headcount): the per-user surface
    (/auth/*, /rules, /execute/:id) was removed. Every endpoint on
    this spec is post-pivot.
  version: 0.2.0
  contact:
    name: LENITNES
servers:
  - url: https://api.lenitnes.io
    description: Production
  - url: http://localhost:8742
    description: Local development (Docker)
  - url: http://localhost:4000
    description: Local development (npm run dev:api, no Docker)

# No global auth — most endpoints are public. Per-endpoint security
# is declared inline (see `security: []` for public routes and
...
security: []
...
paths:
...
  /signals:
    get:
      tags: [Public]
      summary: List signals
      description: |
        Returns recent signals. Heartbeats excluded by default.
        Cached 30s.
      security: []
      parameters:
        - name: monitorId
          in: query
          schema: { type: string, format: uuid }
        - name: includeHeartbeats
          in: query
          schema: { type: boolean, default: false }
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 100 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        '200':
          description: Array of signals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Signal'
...
  /backtest/stats:
    get:
      tags: [Public]
      summary: Backtest statistics per detector + asset
      description: |
        Aggregated accuracy, hit ratio, and avg pct-change per
        detector type and asset.
      security: []
      parameters:
        - name: detector
          in: query
          schema: { type: string }
        - name: asset
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Array of detector stats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DetectorBacktestStats'
...
  /monitors:
    get:
      tags: [Public]
      summary: List watchlist entries
      description: |
        Returns the system watchlist (admin-curated, no per-user
        filter). Cached 30s.
      security: []
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 100 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        '200':
          description: Array of monitors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Monitor'
...
  /orders:
    get:
      tags: [Public]
      summary: List recent orders
      description: Recent treasury trades joined with signal + monitor metadata.
      security: []
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 100 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        '200':
          description: Array of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
  /dlq:
    get:
      tags: [Public]
      summary: Dead-letter queue depth + jobs
      description: BullMQ jobs that exhausted retries. Operator surface in practice.
      security: []
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 200 }
      responses:
        '200':
          description: DLQ snapshot
          content:
            application/json:
              schema:
                type: object
                properties:
                  depth: { type: integer }
                  jobs:
                    type: array
                    items:
                      $ref: '#/components/schemas/DlqJob'
...
components:
  securitySchemes:
    adminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
      description: |
        Matches the `ADMIN_API_KEY` env var. When unset, all
        /admin/* endpoints return 503 (no info leak).
  schemas:
...
    MonitorStatus:
      type: string
      enum: [active, paused, triggered]
    Monitor:
      type: object
      required:
        - id
        - url
        - condition_text
        - frequency_seconds
        - status
        - screenshots_enabled
        - is_public
        - confidence_threshold
        - asset_mapping
        - created_at
      properties:
        id: { type: string, format: uuid }
        url: { type: string, format: uri }
        condition_text: { type: string, maxLength: 500 }
        frequency_seconds: { type: integer, minimum: 60 }
        status: { $ref: '#/components/schemas/MonitorStatus' }
        screenshots_enabled: { type: boolean }
        is_public: { type: boolean }
        confidence_threshold: { type: integer, minimum: 0, maximum: 100 }
        last_check_at: { type: string, format: date-time, nullable: true }
        last_seen_commit_hash: { type: string, nullable: true }
        asset_mapping:
          type: object
          properties:
            coingeckoId: { type: string }
            krakenPair: { type: string }
            tokenizedStock: { type: string }
            direction: { type: string, enum: [long, short, both] }
        created_at: { type: string, format: date-time }
...
    Signal:
      type: object
      required:
        - id
        - monitor_id
        - detected_at
        - is_heartbeat
      properties:
        id: { type: string, format: uuid }
        monitor_id: { type: string, format: uuid }
        detected_at: { type: string, format: date-time }
        hedera_tx_id: { type: string, nullable: true }
        hedera_hcs_message_id: { type: string, nullable: true }
        tinyfish_run_id: { type: string, nullable: true }
        ipfs_cid: { type: string, nullable: true }
        evidence_text: { type: string, nullable: true }
        screenshot_urls:
          type: array
          items: { type: string }
        condition_summary: { type: string, nullable: true }
        is_heartbeat: { type: boolean }
        arb_tx_hash: { type: string, nullable: true }
        orders_count: { type: integer }
...
    Order:
      type: object
      required: [id, signal_id, status]
      properties:
        id: { type: string, format: uuid }
        signal_id: { type: string, format: uuid }
        rule_id:
          type: string
          nullable: true
          description: 'Deprecated — rules table dropped in Day 2 pivot. Always null post-pivot.'
        kraken_order_id: { type: string, nullable: true }
        order_params:
          type: object
          additionalProperties: true
        status:
          type: string
          enum: [pending, placed, filled, partially_filled, cancelled, failed, expired]
        placed_at: { type: string, format: date-time, nullable: true }
        cancelled_at: { type: string, format: date-time, nullable: true }
        kraken_response:
          type: object
          nullable: true
          additionalProperties: true
        chain:
          type: string
          enum: [hedera, arbitrum, robinhood, bnb]
          nullable: true
        chain_tx_hash: { type: string, nullable: true }

    # ── Agent ───────────────────────────────────────────────
...
    DetectorBacktestStats:
      type: object
      required: [detector_type, asset, total_signals, correct_count]
      properties:
        detector_type: { type: string }
        asset: { type: string }
        total_signals: { type: integer }
        correct_count: { type: integer }
        accuracy: { type: string }
        avg_pct_change: { type: string }
        median_pct_change: { type: string }
        avg_abs_return: { type: string }
        sharpe_estimate: { type: string }
        best_window:
          type: integer
          nullable: true
...
    DlqJob:
      type: object
      required: [monitorId, finalError, attemptsMade, movedAt]
      properties:
        monitorId: { type: string }
        finalError: { type: string }
        attemptsMade: { type: integer }
        movedAt: { type: string, format: date-time }

    # ── SoSoValue ─────────────────────────────────────────────
...